adding files

+++ src/screens/Catalog/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import { ScrollView, View, Text, TouchableOpacity, SafeAreaView, Image, AsyncStorage } from 'react-native'
+import { ScrollView, View, Text, TouchableOpacity, SafeAreaView, Alert, AsyncStorage } from 'react-native'
 import Styles from './style'
 import { NavigationEvents } from "react-navigation";
 import { MaterialIcons } from '@expo/vector-icons';
 import AccountListImage from '../../components/AccountList/AccountListImage'
 import { getDataProducts } from '../../config/firebase/Database/GetData'
-import {deleteData} from '../../config/firebase/Database/DeleteData'
+import { deleteData } from '../../config/firebase/Database/DeleteData'
 
 export default class Account extends React.Component {
   constructor() {
@@ -20,26 +20,43 @@ export default class Account extends React.Component {
   async  componentDidMount() {
     const uid = await AsyncStorage.getItem('uid')
     const res = await getDataProducts(uid);
-    this.setState({content: res});    
-    for(let x of res){
+    this.setState({ content: res });
+    for (let x of res) {
       // console.log(x.docData.allData.status)
       const status = x.docData.allData.status
-      if(status === 'pending'){
-        this.setState({pending: this.state.pending + 1})
-      }else if(status === 'rejected'){
-        this.setState({rejected: this.state.rejected + 1})
+      if (status === 'pending') {
+        this.setState({ pending: this.state.pending + 1 })
+      } else if (status === 'rejected') {
+        this.setState({ rejected: this.state.rejected + 1 })
       }
 
     }
   }
 
-  handleDelete = async(docId)=>{
-    await deleteData('product', docId).then((res)=>{
-      console.log(res)
-        if(res === 'sucsses'){
-          this.componentDidMount()
-        }
-    })
+  handleDelete = async (docId) => {
+    Alert.alert(
+      'Are you sure?',
+      'This will delete the product.',
+      [
+        {
+          text: 'Cancel',
+          onPress: () => console.log('Cancel Pressed'),
+          style: 'cancel',
+        },
+        {
+          text: 'OK', onPress:async () => {
+            await deleteData('product', docId).then((res) => {
+              console.log(res)
+              if (res === 'sucsses') {
+                this.componentDidMount()
+              }
+            })
+          }
+        },
+      ],
+      { cancelable: true },
+    );
+
   }
   render() {
     const { loading, content, rejected, pending } = this.state;
@@ -47,12 +64,12 @@ export default class Account extends React.Component {
     return (
 
       <ScrollView style={Styles.container}>
-         <NavigationEvents
-                    onWillFocus={() => {
-                        this.setState({content: [], rejected: 0,pending: 0})
-                        this.componentDidMount()
-                    }}
-                />
+        <NavigationEvents
+          onWillFocus={() => {
+            this.setState({ content: [], rejected: 0, pending: 0 })
+            this.componentDidMount()
+          }}
+        />
         <View style={Styles.mainBox}>
           <TouchableOpacity style={Styles.BoxContainer} onPress={() => { this.props.navigation.navigate("AddProduct") }}>
             <MaterialIcons name="add-box" size={40} color="white" />
@@ -77,20 +94,20 @@ export default class Account extends React.Component {
           </View>
 
           <View>
-            {!!content.length && content.map((item,i)=>{
+            {!!content.length && content.map((item, i) => {
               // console.log(item.docData.allData.allData.allData.data.catData.docId)
-              return(
-                <TouchableOpacity 
-                key={i}
-                style={Styles.catList}
-                onPress={() => { this.handleDelete(item.docId) }}>
-              <AccountListImage
-                title={item.docData.allData.allData.allData.data.title}
-                src={{ uri: item.docData.allData.callBack[0] }}
-                subTitle={"Price $" + item.docData.allData.price}
-                />
-            </TouchableOpacity>
-                )
+              return (
+                <TouchableOpacity
+                  key={i}
+                  style={Styles.catList}
+                  onPress={() => { this.handleDelete(item.docId) }}>
+                  <AccountListImage
+                    title={item.docData.allData.allData.allData.data.title}
+                    src={{ uri: item.docData.allData.callBack[0] }}
+                    subTitle={"Price $" + item.docData.allData.price}
+                  />
+                </TouchableOpacity>
+              )
             })}
           </View>
         </View>
